+++ client/src/App.tsx
@@ -10,7 +10,7 @@ import Road from 'routes/Road';
 import MapDetail from 'components/road/MapDetail';
 import Meetups from 'routes/Meetups';
 import Meetup from 'routes/Meetup';
-import Login from 'routes/AuthPage/Login';
+import Login from 'routes/Auth/Login';
 import Toast from 'components/common/Toast';
 // import ErrorBoundary from 'components/ErrorBoundary';
 import { createTheme } from '@mui/material';
@@ -19,12 +19,12 @@ import { MAIN_COLOR } from 'utils/constants';
 import { Theme as MuiTheme } from '@mui/material/styles';
 import PrivateRoute from 'components/common/PrivateRoute';
 import Search from 'routes/Search';
-import AuthPage from 'routes/AuthPage';
+import AuthPage from 'routes/Auth';
 import PublicOnlyRoute from 'components/common/PublicOnlyRoute';
 import Modal from 'components/common/Modal';
-import Signup from 'routes/AuthPage/Signup';
-import FindEmail from 'routes/AuthPage/FindEmail';
-import ResetPassword from 'routes/AuthPage/ResetPassword';
+import Signup from 'routes/Auth/Signup';
+import FindEmail from 'routes/Auth/FindEmail';
+import ResetPassword from 'routes/Auth/ResetPassword';
 import MyPage from 'routes/Mypage';
 import MyMeetups from 'routes/Mypage/MyMeetups';
 import JoinedMeetups from 'routes/Mypage/JoinedMeetups';
