+++ client/src/components/meetups/MeetupInfo/index.tsx
@@ -2,15 +2,14 @@ import {
   stringifyMeetupPathDifficulty,
   stringifyRidingSkill,
 } from 'utils/stringify';
-import styles from './MeetupSummary.module.scss';
+import styles from './MeetupInfo.module.scss';
 import classNames from 'classnames/bind';
 import dayjs from 'dayjs';
 import { MeetupData } from 'types/meetup';
 
 const cn = classNames.bind(styles);
 
-// TODO: rename
-const MeetupSummary = (props: MeetupData) => (
+const MeetupInfo = (props: MeetupData) => (
   <div className={cn('container')}>
     <div className={cn('text')}>
       <header className={cn('header')}>
@@ -71,4 +70,4 @@ const MeetupSummary = (props: MeetupData) => (
   </div>
 );
 
-export default MeetupSummary;
+export default MeetupInfo;
