+++ client/src/components/meetups/MeetupItem/index.tsx
@@ -1,8 +1,8 @@
 import { memo } from 'react';
 import { MeetupData } from 'types/meetup';
 import { Link } from 'react-router-dom';
-import MeetupSummary from '../MeetupSummary';
-import MeetupRoutePreview from '../MeetupRoutePreview';
+import MeetupInfo from '../MeetupInfo';
+import MeetupRoute from '../MeetupRoute';
 import styles from './MeetupItem.module.scss';
 import classNames from 'classnames/bind';
 
@@ -15,8 +15,8 @@ interface MeetupItemProp {
 const MeetupItem = memo(({ meetup }: MeetupItemProp) => (
   <Link to={`/meetups/${meetup.id}`} className={cn('link')}>
     <article className={cn('meetup')}>
-      <MeetupSummary {...meetup} />
-      <MeetupRoutePreview courseName={meetup.courseName} path={meetup.path} />
+      <MeetupInfo {...meetup} />
+      <MeetupRoute courseName={meetup.courseName} path={meetup.path} />
     </article>
   </Link>
 ));
