+++ client/src/routes/Auth/index.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import styles from './AuthPage.module.scss';
+import styles from './Auth.module.scss';
 
 const AuthPage = () => (
   <section className={styles.auth}>
