+++ client/src/routes/Meetups/index.tsx
@@ -44,9 +44,9 @@ const Meetups = () => {
     {
       // TODO: order 디폴트 null?
       select: meetups => getMeetupsOrderedBy(order.name, meetups),
-      keepPreviousData: true,
-      staleTime: 60 * 1000,
+      staleTime: 5 * 60 * 1000,
       cacheTime: Infinity,
+      keepPreviousData: true,
       onError: () => showToastMessage('로딩 중 문제가 발생했습니다.'),
     }
   );
